+++ changelog.md
@@ -1,5 +1,6 @@
 ## v0.0.94
 * EventSchemaLoader uses JsonParser for event data rather than YAMLParser - T227484
+* EventSparkSchemaLoader now merges input JSONSchema with Hive schema before loading - T227088
 
 ## v0.0.93
 * Refactor mediawiki-page-history computation + fix